# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is for demonstration purposes only.",
        "title": "Config Manager APIs",
        "termsOfService": "http://github.com/gofreego/configo/readme.md",
        "contact": {
            "name": "Developers",
            "url": "http://www.github.com/gofreego",
            "email": "pavanyewale1996@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/configo/config": {
            "post": {
                "description": "Save config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Save config",
                "parameters": [
                    {
                        "description": "config object",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configo.UpdateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "config saved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/configo/metadata": {
            "get": {
                "description": "Get all config keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Get all config keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configo.configMetadataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/configo/web/": {
            "get": {
                "description": "UI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "UI",
                "responses": {
                    "200": {
                        "description": "UI",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/configs/confio/{key}": {
            "get": {
                "description": "Get config by key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Get config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "config key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configo.ConfigObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "configo.ConfigInfo": {
            "type": "object",
            "properties": {
                "configKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "configo.ConfigObject": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configo.ConfigObject"
                    }
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/configo.ConfigType"
                },
                "value": {}
            }
        },
        "configo.ConfigType": {
            "type": "string",
            "enum": [
                "string",
                "number",
                "boolean",
                "json",
                "big_text",
                "choice",
                "parent",
                "list"
            ],
            "x-enum-varnames": [
                "CONFIG_TYPE_STRING",
                "CONFIG_TYPE_NUMBER",
                "CONFIG_TYPE_BOOLEAN",
                "CONFIG_TYPE_JSON",
                "CONFIG_TYPE_BIG_TEXT",
                "CONFIG_TYPE_CHOICE",
                "CONFIG_TYPE_PARENT",
                "CONFIG_TYPE_LIST"
            ]
        },
        "configo.ServiceInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "configo.UpdateConfigRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/configo.ConfigObject"
                }
            }
        },
        "configo.configMetadataResponse": {
            "type": "object",
            "properties": {
                "configInfo": {
                    "$ref": "#/definitions/configo.ConfigInfo"
                },
                "serviceInfo": {
                    "$ref": "#/definitions/configo.ServiceInfo"
                }
            }
        }
    }
}